requirements.txt
flask
openai
psycopg2-binary
python-dotenv

Flask
flask-cors
python-dotenv
openai
